+++ tests/test_live_client.py
@@ -5,6 +5,7 @@ Unit tests for the Live client.
 from __future__ import annotations
 
 import pathlib
+import platform
 import random
 import string
 from io import BytesIO
@@ -503,6 +504,51 @@ async def test_live_subscribe_large_symbol_list(
     assert reconstructed == large_symbol_list
 
 
+async def test_live_subscribe_from_callback(
+    live_client: client.Live,
+    mock_live_server: MockLiveServer,
+) -> None:
+    """
+    Test that `Live.subscribe` can be called from a callback.
+    """
+    # Arrange
+    live_client.subscribe(
+        dataset=Dataset.GLBX_MDP3,
+        schema=Schema.OHLCV_1H,
+        stype_in=SType.RAW_SYMBOL,
+        symbols="TEST0",
+    )
+
+    def cb_sub(_: DBNRecord) -> None:
+        live_client.subscribe(
+            dataset=Dataset.GLBX_MDP3,
+            schema=Schema.MBO,
+            stype_in=SType.RAW_SYMBOL,
+            symbols="TEST1",
+        )
+
+    live_client.add_callback(cb_sub)
+
+    # Act
+    first_sub = mock_live_server.get_message_of_type(
+        gateway.SubscriptionRequest,
+        timeout=1,
+    )
+
+    live_client.start()
+
+    await live_client.wait_for_close()
+
+    second_sub = mock_live_server.get_message_of_type(
+        gateway.SubscriptionRequest,
+        timeout=1,
+    )
+
+    # Assert
+    assert first_sub.symbols == "TEST0"
+    assert second_sub.symbols == "TEST1"
+
+
 @pytest.mark.usefixtures("mock_live_server")
 def test_live_stop(
     live_client: client.Live,
@@ -794,6 +840,7 @@ def test_live_add_stream_path_directory(
         live_client.add_stream(tmp_path)
 
 
+@pytest.mark.skipif(platform.system() == "Windows", reason="flaky on windows")
 async def test_live_async_iteration(
     live_client: client.Live,
 ) -> None:
@@ -910,6 +957,7 @@ async def test_live_async_iteration_dropped(
     assert live_client._dbn_queue.empty()
 
 
+@pytest.mark.skipif(platform.system() == "Windows", reason="flaky on windows")
 async def test_live_async_iteration_stop(
     live_client: client.Live,
 ) -> None:
@@ -936,6 +984,7 @@ async def test_live_async_iteration_stop(
     assert live_client._dbn_queue.empty()
 
 
+@pytest.mark.skipif(platform.system() == "Windows", reason="flaky on windows")
 def test_live_sync_iteration(
     live_client: client.Live,
 ) -> None: