+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.31.1 - 2024-03-20
+
+#### Enhancements
+- Increase `Live` session connection and authentication timeouts
+- Added new `F_TOB` and `F_MAYBE_BAD_BOOK` variants to `RecordFlags`
+
+#### Bug fixes
+- Fixed an issue where calling `Live.subscribe` from a `Live` client callback would cause a deadlock
+
 ## 0.31.0 - 2024-03-05
 
 #### Enhancements